requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
elastic-transport==8.4.1
elasticsearch==8.10.0
elasticsearch-dsl==8.9.0
exceptiongroup==1.1.3
fastapi==0.103.1
frozenlist==1.4.0
h11==0.14.0
hazelcast-python-client==5.3.0
httptools==0.6.0
idna==3.4
iniconfig==2.0.0
motor==3.3.1
multidict==6.0.4
packaging==23.2
pluggy==1.3.0
pydantic==2.3.0
pydantic_core==2.6.3
pymemcache==4.0.0
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.7.1
urllib3==1.26.17
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
yarl==1.9.2